UT new Unit test to validate Quote request
